scrapper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ibtikar-chatbot"
version = "0.1.0"
description = "Arabic-first RAG chatbot system for Ibtikar community"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Ibtikar Development Team", email = "dev@ibtikar.org"}
]
keywords = ["chatbot", "RAG", "Arabic", "NLP", "FastAPI", "LangChain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Education",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "langchain>=0.0.350",
    "langchain-community>=0.0.10",
    "sentence-transformers>=2.2.2",
    "chromadb>=0.4.18",
    "faiss-cpu>=1.7.4",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "prometheus-client>=0.19.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "selenium>=4.15.0",
    "playwright>=1.40.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "azure-storage-blob>=12.19.0",
    "upstash-vector>=0.4.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
gpu = [
    "faiss-gpu>=1.7.4",
    "torch>=2.1.0",
]

[project.urls]
Homepage = "https://github.com/ibtikar-org-tr/ibtikar-chatbot"
Documentation = "https://github.com/ibtikar-org-tr/ibtikar-chatbot/wiki"
Repository = "https://github.com/ibtikar-org-tr/ibtikar-chatbot.git"
Issues = "https://github.com/ibtikar-org-tr/ibtikar-chatbot/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ibtikar_chatbot"]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
